Code cleanups (courtesy of IntelliJ Idea code inspector)
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpcore/trunk@1487830 13f79535-47bb-0310-9956-ffa450edef68

+++ httpcore/src/main/java/org/apache/http/pool/AbstractConnPool.java
@@ -319,7 +319,7 @@ public abstract class AbstractConnPool<T, C, E extends PoolEntry<T, C>>
     private int getMax(final T route) {
         final Integer v = this.maxPerRoute.get(route);
         if (v != null) {
-            return v.intValue();
+            return v;
         } else {
             return this.defaultMaxPerRoute;
         }
